chore: move/refactor solari-transfers valinor code
