+++ solari-transfers/src/valinor/NOTICE
