+++ solari-transfers/src/valinor/VALINOR_LICENSE
