+++ solari-transfers/src/valinor/edge_models.rs
