--- solari-transfers/src/valinor/export_edges/mod.rs
@@ -1,2 +1,0 @@
-pub mod edge_export;
-mod edge_models;
